requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==4.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
Flask==2.3.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
multidict==6.7.0
openai==0.28.0
propcache==0.4.1
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.5
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tornado==6.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
